pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0.0",
    "setuptools-scm>=8.0",
]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
name = "pdi-nomad-plugin"
dynamic = ["version"]
description = "A plugin for PDI NOMAD containing principal techniques."
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Andrea Albino", email = "andrea.albino@physik.hu-berlin.de" },
    { name = "Hampus Näsström", email = "hampus.naesstroem@physik.hu-berlin.de" },
    { name = "Sarthak Kapoor", email = "sarthak.kapoor@physik.hu-berlin.de" },
    { name = "Sebastian Brückner", email = "sebastian.brueckner@ikz-berlin.de" },
]
maintainers = [
    { name = "Andrea Albino", email = "andrea.albino@physik.hu-berlin.de" },
]
license = { file = "LICENSE" }
dependencies = [
    'nomad-lab>=1.3.1',
    'nomad-material-processing',
    'nomad-measurements',
    'nomad-analysis',
    'lakeshore-nomad-plugin@git+https://github.com/IKZ-Berlin/lakeshore-nomad-plugin.git@main',
    'epic_scraper@git+https://github.com/PDI-Berlin/EPIC_scraper@main',
]
[project.urls]
"Homepage" = "https://github.com/PDI-Berlin/pdi-nomad-plugin"
"Bug Tracker" = "https://github.com/PDI-Berlin/pdi-nomad-plugin/issues"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']

general_schema = "pdi_nomad_plugin.general:schema"
characterization_schema = "pdi_nomad_plugin.characterization:schema"
materials_schema = "pdi_nomad_plugin.mbe:materials_schema"
instrument_schema = "pdi_nomad_plugin.mbe:instrument_schema"
processes_schema = "pdi_nomad_plugin.mbe:processes_schema"
# epic_mbe_parser = "pdi_nomad_plugin.mbe.epic_parser:epic_parser"
config_mbe_parser = "pdi_nomad_plugin.mbe.epic_parser:config_parser"
mbe_app = "pdi_nomad_plugin.mbe.mbe_app:substrateapp"